requirements.txt
websockets>=11.0.3
requests>=2.31.0
asyncio
python-dotenv
pycryptodome